Add files via upload
In this project, my teammate and I, developed a GUI where we wanted to simulate an application for acquaring a parking lot. In other words, we tried to manage the available places in the parking of the university by taking control of who and when was he / she trying to obtain a parking lot. This project was made in first semester of college.